+++ packages/registry/src/index.ts
@@ -11,3 +11,4 @@ export * from "./tokens";
 export * from "./chains";
 export * from "./assets";
 export * from "./coverage";
+export * from "./drip";